init: reset settings
