+++ week3/src/App.css
@@ -36,7 +36,3 @@
     transform: rotate(360deg);
   }
 }
-
-button {
-  font-size: calc(10px + 2vmin);
-}
